Adjustments to model
